+++ Source/Rect.h
@@ -11,7 +11,7 @@ private:
 	Uint8 b;
 	Uint8 a;
 public:
-	Rect( int x, int y, int width, int height, Uint8 r = 255, Uint8 g = 0, Uint8 b = 0, Uint8 a = 255 );
+	Rect( float x, float y, float width, float height, Uint8 r = 255, Uint8 g = 0, Uint8 b = 0, Uint8 a = 255 );
 	~Rect();
 
 	void setColor( Uint8 r, Uint8 g, Uint8 b, Uint8 a);
